+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Release Notes
 
 - 2.4.3 *Unreleased*
-    - Fix font-size slider’s colouring.
+    - Fix font-size slider control’s colouring.
 - 2.4.2 *16 April 2026*
     - Fix paragraph breaks in content copied from previews.
     - Reduce image placeholder sizes and change referenced path's colour.